requirements.txt
anthropic>=0.40.0
clickhouse-connect>=0.7.0
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tabulate>=0.9.0
python-multipart>=0.0.6
